+++ mayan/apps/linking/apps.py
@@ -11,8 +11,13 @@ from mayan.apps.common.html_widgets import TwoStateWidget
 from mayan.apps.common.menus import (
     menu_facet, menu_list_facet, menu_object, menu_secondary, menu_setup
 )
+from mayan.apps.events.classes import ModelEventType
+from mayan.apps.events.links import (
+    link_events_for_object, link_object_event_types_user_subcriptions_list
+)
 from mayan.apps.navigation.classes import SourceColumn
 
+from .events import event_smart_link_created, event_smart_link_edited
 from .links import (
     link_smart_link_create, link_smart_link_condition_create,
     link_smart_link_condition_delete, link_smart_link_condition_edit,
@@ -37,6 +42,7 @@ class LinkingApp(MayanAppConfig):
 
     def ready(self):
         super(LinkingApp, self).ready()
+        from actstream import registry
 
         Document = apps.get_model(
             app_label='documents', model_name='Document'
@@ -45,6 +51,12 @@ class LinkingApp(MayanAppConfig):
         ResolvedSmartLink = self.get_model(model_name='ResolvedSmartLink')
         SmartLink = self.get_model(model_name='SmartLink')
         SmartLinkCondition = self.get_model(model_name='SmartLinkCondition')
+
+        ModelEventType.register(
+            event_types=(
+                event_smart_link_created, event_smart_link_edited
+            ), model=SmartLink
+        )
 
         ModelPermission.register(
             model=SmartLink, permissions=(
@@ -79,7 +91,9 @@ class LinkingApp(MayanAppConfig):
         )
         menu_list_facet.bind_links(
             links=(
-                link_acl_list, link_smart_link_document_types,
+                link_acl_list, link_events_for_object,
+                link_smart_link_document_types,
+                link_object_event_types_user_subcriptions_list,
                 link_smart_link_condition_list,
             ), sources=(SmartLink,)
         )
@@ -115,3 +129,5 @@ class LinkingApp(MayanAppConfig):
             )
         )
         menu_setup.bind_links(links=(link_smart_link_setup,))
+
+        registry.register(SmartLink)