+++ HISTORY.rst
@@ -194,7 +194,10 @@
   from the AddRemoveView.
 * Add view to link document type and indexes from the document
   type side.
-
+* Update smart link document type selection view to use
+  AddRemoveView class.
+* Add smart link created and edited events.
+  
 3.1.11 (2019-04-XX)
 ===================
 * Fix multiple tag selection wizard step.